Add files via upload
I'm almost done with my portion! All that is left is to figure out how to compare class to the rest of the variables, but I've outlined my general idea in the comments and the code for the actual regression is there and working. Please look it over and let me know if there are any changes that need to be made or if there is an easier way to incorporate class. or if I even need to tbh. Thanks!